order-service/requirements.txt
Flask==2.3.2
pymongo==4.5.0
pika==1.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
